Step 2: Rewrite update to target the correct API endpoint
Also add some error handling along the way - this makes sure that the updates work.

Also aborted the S3 idea (High costs) and R2 idea (Not available in India coz of payment issues)

+++ board/raspberrypi/rootfs-overlay/etc/init.d/update.sh
@@ -1,13 +1,15 @@
 #!/bin/sh
+curl --max-time 240 -L -k -o /tmp/suvriddhi.zip https://github.com/raviishh/SuvriddhiOS/releases/latest/download/suvriddhi.zip
+if [ $? -ne 0 ]; then
+  echo "Download failed, aborting update"
+  exit 1
+fi
 
-# Step 1: Curl the release URL:
-curl --max-time 30 -k -o /tmp/release.json https://api.github.com/repos/raviishh/SuvriddhiOS/releases/latest
-DOWNLOAD_URL=$(grep -o '"browser_download_url": *"[^"]*"' /tmp/release.json | grep -o 'https://[^"]*')
-curl --max-time 60 -L -k -o /tmp/suvriddhi.zip "$DOWNLOAD_URL"
-
-# Step 2: After downloading you have a zip file, create /root/www_new with the extracted contents of the zip
 mkdir -p /root/www_new
 unzip -o /tmp/suvriddhi.zip -d /root/www_new
+if [ $? -ne 0 ]; then
+  echo "Unzip failed, aborting update"
+  exit 1
+fi
 
-# Step 3: Reboot and have S55git take over the rest on boot.
 reboot
